added windows support

+++ app/cli/GuiAppStarter.java
@@ -6,22 +6,29 @@ import java.io.File;
 import conf.PathConfiguration;
 
 public final class GuiAppStarter {
+	private final static String OS;
+	public static final String SERVER_EXECUTABLE = PathConfiguration.projectRoot + "target/universal/stage/bin/"; 
+	private final static String START_SCRIPT = "./premier-league";
+	private final static String SERVER_PID = PathConfiguration.projectRoot + "target/universal/stage/";
+	private final static String APP_PORT = "http://localhost:9000/app/premier-league";
 	private static ProcessBuilder processBuilder;
 	private static Process serverProcess;
 	private static Process appProcess;
-	public static final String projectFinalBuildExecutablePath = PathConfiguration.projectRoot + "target/universal/stage/bin/"; 
-	private final static String startScript = "./premier-league";
-	private final static String serverPIDPath = PathConfiguration.projectRoot + "target/universal/stage/";
-	private final static String appPort = "http://localhost:9000/app/premier-league";
-
+	static {
+		OS = System.getProperty("os.name").toLowerCase();
+	}
 	private GuiAppStarter() {}
 
 	public static void start() {
 		removeAndStopPID();
 		if(serverProcess == null) {
 			processBuilder = new ProcessBuilder();
-			processBuilder.directory(new File(projectFinalBuildExecutablePath));
-			processBuilder.command("bash", "-c", startScript);
+			processBuilder.directory(new File(SERVER_EXECUTABLE));
+			if (isWindows()) {
+				processBuilder.command("cmd.exe", "/c", START_SCRIPT + ".bat");
+			} else {
+				processBuilder.command("bash", "-c", START_SCRIPT);
+			}
 			try {
 				serverProcess = processBuilder.start();
 			} catch (Exception e) {
@@ -53,11 +60,19 @@ public final class GuiAppStarter {
 	}
 	private static void removeAndStopPID() {
 		processBuilder = new ProcessBuilder();
-		processBuilder.directory(new File(serverPIDPath));
+		processBuilder.directory(new File(SERVER_PID));
 		try {
-			processBuilder.command("bash", "-c", "kill `cat RUNNING_PID`");
+			if(isWindows()) {
+				processBuilder.command("cmd.exe", "/c", "taskkill /F /PID `type RUNNING_PID`");
+	 		} else {
+				processBuilder.command("bash", "-c", "kill `cat RUNNING_PID`");
+			}
 			processBuilder.start();
-			processBuilder.command("bash", "-c", "rm RUNNING_PID");
+			if(isWindows()) {
+				processBuilder.command("cmd.exe", "/c", "del RUNNING_PID");
+			} else {
+				processBuilder.command("bash", "-c", "rm RUNNING_PID");
+			}
 			processBuilder.start();
 		} catch(Exception e) {
 			e.printStackTrace();
@@ -79,7 +94,11 @@ public final class GuiAppStarter {
 		// return;
 		// } 
 		processBuilder = new ProcessBuilder();
-		processBuilder.command("bash", "-c", "google-chrome-stable --app=" + appPort);
+		if(isWindows()) {
+			processBuilder.command("cmd.exe", "/c", "start chrome --app=" + APP_PORT);
+		} else {
+			processBuilder.command("bash", "-c", "google-chrome-stable --app=" + APP_PORT);
+		}
 		try {
 			appProcess = processBuilder.start();
 		} catch (Exception e) {
@@ -88,11 +107,13 @@ public final class GuiAppStarter {
 		}
 	}
 
-
+	private static boolean isWindows() {
+		return OS.contains("windows");
+	}
 	@Deprecated
 	public static void openApp() {
 		try {
-			URI uri = new URI(appPort);
+			URI uri = new URI(APP_PORT);
 			if (Desktop.isDesktopSupported() && Desktop.getDesktop().isSupported(Desktop.Action.BROWSE)) {
 				try {
 					Desktop.getDesktop().browse(uri);
